continue adding support for ELF64 read & write
We are getting really good object files now. I think only the relocations are
missing, which should be the next thing to be added.

There are other changes that could be made, such as removing unused code from .o files.
Right now, we are copying the entire .text section from the original binary into each .o
file. This is the naive approach, as it is a superset of what the .o really references,
and a bit wasteful.

Signed-off-by: Joel Nider <joeln@il.ibm.com>