Add ll_remove() to remove a node from a linked list
This one is a bit complicated, because it deals with the opaque data. We need
to read the opaque data in order to know that we have found the correct node
that the user wants to remove. So we ask the user to provide a comparator function
that will return 0 when we found the correct node. Then we iterate through all
nodes, calling the user-provided function. When we find the (first/right) one,
we remove the node from the linked list, and return a pointer to the node to the
user so it can be deleted or whatever.

Signed-off-by: Joel Nider <joeln@il.ibm.com>