Remove index from backend section
Having a stored index number is causing problems. We have to know the index in order
to link sections together when writing a ELF object file, but the index must be
according to the order in which the sections are written in the file. Originally we
wrote all sections including potentially empty sections in order to have a static
ordering of sections. This fell apart at some point, and also including useless
information in output files. So now we are fully dynamic, and section numbers are
calculated when the file is being written.

Signed-off-by: Joel Nider <joeln@il.ibm.com>