Write ELF sections dynamically
That means we no longer have static ordering of sections - sections are written
as they appear in the linked list of the backend object. That also means all
sections must exist (some are created on-the-fly in the write_file function)
including the string tables and symbol tables. Links between the sections are
also generated when the file is written out. This means only sections that are
in the linked list will be written out, making the files a bit more compact
and easier to debug since we have less extraneous information.

Signed-off-by: Joel Nider <joeln@il.ibm.com>