backend: Add new symbol accessor functions
Add 3 new functions for symbol manipulation: 2 for search and 1 for insertion.

The function backend_find_symbol_by_val_type() avoids the assumption that there
is only 1 symbol per address, and allows selecting a symbol by type in the case
that there is more than one.

The funciont backend_find_nearest_symbol() relaxes the assumption that a symbol
exists at a particular address, and returns the previous symbol that contains
the searched address.

The function backend_split_symbol allows us to split an existing symbol at a
particular point into two symbols of the same kind.

Signed-off-by: Joel Nider <joel@nider.org>