fix segfault when adding function with no file
If a backend object is not opened, we cannot add symbols to it.
In the case that there is no 'file' symbol, or a 'function' symbol
is encountered before the 'file' symbol, the backend object will
not yet be created, leading to a NULL pointer being passed to the
add function.

Signed-off-by: Joel Nider <joel@nider.org>