Add support for DOS EXE format
This format is also known as 'MZ' because of the signature. This is
16-bit x86 real mode code used in the DOS operating system. It is a
much simpler format, but also doesn't really follow modern
conventions (like separate sections for code & data) which presents
some challenges for reverse engineering automatically.

It is likely that 'nucleus' function detection will do a better job
with this kind of a file, since it builds an execution tree rather
than my own brute force 'disassemble everything' approach.

Signed-off-by: Joel Nider <joel@nider.org>