Prepare for move to C++
There are 2 main reasons for moving to C++.
1) to standardize parts of the code (such as the linked list implementation)
2) to use Nucleus without major changes

As the first step in the move to C++, I compiled the code with g++ rather
than gcc. This commit fixes all of the errors & warnings that were found
as part of that exercise. However, I did not update the makefile, meaning
gcc is still the compiler is use.

Signed-off-by: Joel Nider <joel@nider.org>