PE: Implement generic handling of import tables
The current implementation of the PE32 backend is not able to handle
certain cases of the import tables in certain sections. This patch
makes the handling more generic, so it can handle more cases.

It also uses data that has already be loaded into memory, which may
be faster than reading 4 bytes at a time from a file.

Signed-off-by: Joel Nider <joel@nider.org>