Add wrapper for 'nucleus'
The nucleus backend format is different from our backend format.
It some ways it is better (uses C++ list for example) and in some
ways it is inferior (does not hold as much information detail, and
the API is not as rich). In any case, it is necessary to convert
to the nucleus backend format in order to call their functions for
reconstructing the symbol table.

Signed-off-by: Joel Nider <joel@nider.org>