Fix problem with fread return value check
This commit could have been avoided. I didn't read the man page
carefully enough, and didn't test the code before commiting the
previous change which checks for the fread() return value. Instead
of comparing it to the number of objects read, I tried to compare
it to the number of bytes read, which fails every time.

Signed-off-by: Joel Nider <joel@nider.org>