Add backend section type
A section can have a type, which is more useful for distinguishing it
than its name. There may be many sections with different names but
that share a type, and should be processed in the same manner. This
helps us remove ambiguity and make the code more generic by not having
to depend on a specific section name (which may or may not exist).

Signed-off-by: Joel Nider <joel@nider.org>