backend: check range in find_symbol_by_val()
Rather than have find_symbol_by_range() only check for a specific starting
value of the function, have it check that the target address falls anywhere
inside the function (i.e. from function start to start+length).
This allows me to check a target function that is covering a relocation,
since relocations are not restricted to the first instruction of a function.

Signed-off-by: Joel Nider <joel@nider.org>