backend: implement backend_sort_symbols() properly
This is now required when writing an ELF object file. The ELF spec
says there is an order in which symbols must appear in the file, so
we now sort these files in the backend before writing the object.

This function is generic (not ELF specific). In order to implement
ordering required by a particular backend file format, a user-defined
comparator function is passed in, and used during sort. I think this
idea is pretty standard by now, and quite useful.

Signed-off-by: Joel Nider <joel@nider.org>