elf: implement the elfcmp function
The elfcmp function is used when the symbols are being sorted before
being written to a file. The symbols must be in a particular order,
and this function defines that ordering. The generic backend performs
the sort by calling this function repeatedly when comparing pairs
of symbols during the sorting algorithm.

Signed-off-by: Joel Nider <joel@nider.org>