elf: 2 bugs with 'flags'
Bug 1: sh.flags was not initialized for .strtab and .shstrtab sections.
That variable inherited a value from another section which has the flag
set to require an index to another section in the sh.info field. Since
this index was not present, other programs (readelf) reported an error
that it was missing, while objdump simply refused to process the file.

Bug 2: sh.info is divided into two fields (in some sections): type and
scope. When reading this flag, I used a macro ELF_SYM_SCOPE, which did
a shift and a mask, and gave me a nice value. Unfortunately, I was
comparing this to a #define which had the mask without the shift,
thereby not recognizing this flag ever (which is the 'global' bit on
symbols). To make things simple, I just don't use the macro anymore.

Signed-off-by: Joel Nider <joel@nider.org>