backend: always add new symbols to the end
I was trying to do something clever, which backfired (as usual). When
adding a 'section' symbol, I tried to make sure they were at the
beginning of the list by pushing them at the head, whereas all other
symbols were added at the tail. This has the side effect of reversing
the order of all of the section symbols, which led to horrible
problems when trying to find the right one by index.

Signed-off-by: Joel Nider <joel@nider.org>