delinker: handle long filenames correctly
A subtle bug - when a long filename is followed by an even longer filename,
the NULL is not being set correctly, leading to a buffer overflow on the
following strcat. Fixed by always doing a memset before writing the filename.
Also, the strncat was replaced with strcat since the src is always .o - a
known length which means strncat doesn't provide me with any benefit.

Signed-off-by: Joel Nider <joel@nider.org>