elf: extensible section header string table
Sometimes we run out of space in the section header string table.
This is possibly the result of a bug somewhere else (i.e. too much
information is being added to the string table to begin with) but
it doesn't hurt to have a flexible mechanism in place in case it
is really needed.

Signed-off-by: Joel Nider <joel@nider.org>