backend: include filename inside object
The backend object now includes the target filename. This is slightly
more flexible because the logic for creating the file can be separated
from the function to write the file. We can now access the filename
for the object that is being written at various stages, without having
to keep an external variable.

Signed-off-by: Joel Nider <joel@nider.org>