delinker: support for multiple symbols per object file
Each source file creates one or more symbols. It is advantageous to
restore the original structure as closely as possible, by writing
all of the symbols to the .o file that were created by the original
source file.

This is the first step - there is still some debugging left to do.