Add 'ignore' list for symbols
If these symbols are encountered while writing to a new backend, they
will be ignored. That means they are not copied into the new backend.
The symbols to be ignored can be added on the command line with -I.
The -I flag can be used multiple times, with one symbol per flag.
Relocations referencing the ignored symbols are created, as well as
relocations that are used by the ignored symbols.

Signed-off-by: Joel Nider <joel@nider.org>