delinker: copy .rodata to output file
When there are data relocations that are based on the section, we must
copy the entire section contents, since we don't know how much data is
going to be used. We must make sure that the output section exists,
and that the buffer is large enough to accomodate the symbols.

Signed-off-by: Joel Nider <joel@nider.org>