elf: look up string table by type
When parsing the ELF file, we should not depend on the string table
to have a particular name. While the name "strtab" is most often used
by convention, it does not have to have this name. Each section has a
type associated with it as well as a name. We should use the type to
identify the section, rather than the name.

Signed-off-by: Joel Nider <joel@nider.org>