backend: remove CODE flag
Some backends have a flag for a code section but also have flags for
attributes such as executable and read/write. I don't really understand
why we need both in the backend, and in fact it leads to some
misunderstandings about what the contents of the section are for.

I'm removing the 'CODE' flag, and replacing it with 'executable' where
necessary. This may impact the generation (output) of some formats
where the CODE flag is used explicitly. I believe I have restored the
correct behaviour, but didn't test it yet.

Signed-off-by: Joel Nider <joel@nider.org>