elf: Fix decoding of PLT entry for x86_64
Now we are starting to get into more details - PLT entries are not
the same on all platforms. We must have different decoding of PLT
entries for each platform because of these differences. I have added
a function for decoding x86_64 PLT entries which include the ENDBR64
instruction (jump target). A similar function must be added for other
platforms in the future.

Signed-off-by: Joel Nider <joel@nider.org>