elf: Add an import function from a relocation
Some relocations point from an intermediate jump table (like the PLT)
to the final destination (like the GOT). We need to be able to extract
the symbol from the relocation in RELA format in a few places. This
generic function can be used on .rela.dyn and .rela.plt for that
purpose.

Signed-off-by: Joel Nider <joel@nider.org>