elf: Handle dynamic symbols properly
Proper handling of dynamic symbols includes reading the .dynamic
section (containing the DT tags) to find all of the imported symbols.
The imported symbols are needed to map 'call' instructions in the
text section to their actual symbols that reside in the GOT. We need
to remove this level of abstraction (the PLT) so it appears more
closely to what a compiler would produce.

Signed-off-by: Joel Nider <joel@nider.org>