delinker: create relocations in a more generic way
Rather than depending on section names, the relocation creation should
be more generic and rely on the attributes (flags) of the sections
instead. Combined with the updated symbol recreation method and PLT
decoding in the ELF backend, this is a much more robust and reliable
implementation for x86 ELF64. In fact, in my (limited) testing, there
are no longer any missing symbols from the .text section -- all
missing symbols are data or debugging symbols (which is expected).

Signed-off-by: Joel Nider <joel@nider.org>