delinker: fix up data relocations
For various sections, the rules are different. The .bss section is not
a PROGBITS section; it is NOBITS. Other than that, it acts like an
uninit data section. The .rel.ro section is of course, a writable,
read-only section. And .rodata is a boring old read-only section.
All of the relocs for these sections are based off the section symbol
which points to the section's base address. Plus, they are all PC-
relavtive (PCREL) but without the -4 needed for functions.

This produces code (.o) that cannot support PIE (position-independent
executable) linking, which is now standard on many systems.

Signed-off-by: Joel Nider <joel@nider.org>