mz: add .data section
Originally I thought I could get away with modeling an MZ file with
a single section that combines code and data together - I was wrong.
When creating relocations, some functions don't exist because the
function reconstruction is not perfect. I know these are functions
because they come from 'call' instructions, and nobody calls data.
But if the function symbol doesn't exist, the code assumes this is a
data symbol, which is perfectly valid since the data and code cover
the same range of addresses. If I separate them, I can detect that
I am trying to relocate into the middle of an existing function, and
that I should split the function. But all of this relies on having
separate .text and .data sections to start with.

Signed-off-by: Joel Nider <joel@nider.org>