Add option to set entry point name
The entry point is often either 'main' or '_start'. The problem is I
don't know which one will be needed. So I will give the user ultimate
control over the selecting the name, by setting it to an arbitrary
string from the command line. If the user does not provide one, it
will default to 'main'.

Signed-off-by: Joel Nider <joel@nider.org>